Added called of verification if the current user can make the change

+++ src/com/sap/controller/AdminController.java
@@ -3,6 +3,8 @@ package com.sap.controller;
 
 import com.sap.model.User;
 import com.sap.service.UserService;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -20,15 +22,18 @@ public class AdminController {
 
     @RequestMapping(value = "/adminPageTest")
     public String listUsers (Model model){
+
+        User admin = userService.getCurrentUser();
+        
         model.addAttribute("user", new User());
-        model.addAttribute("users", userService.getAllUsers());
+        model.addAttribute("users", userService.getTeamMates(admin));
         return "adminPageTest";
     }
 
     @RequestMapping(value = "/user/add", method = RequestMethod.POST)
     public String addUser (@ModelAttribute("user") User user){
 
-        userService.saveUser(user);
+        userService.saveUserAtOwnerTeam(user, userService.getCurrentUser());
 
         return "redirect:/adminPageTest";
 
@@ -37,16 +42,31 @@ public class AdminController {
     @RequestMapping(value = "remove/{id}")
     public String removeUser (@PathVariable("id") Integer id){
 
-        userService.deleteUserByID(id);
+        if (userService.verifyIfActionCanBeAppliedToUser(userService.getUserByID(id))) {
 
+            userService.deleteUserByID(id);
+
+        }
         return "redirect:/adminPageTest";
     }
 
     @RequestMapping(value = "/edit/{id}")
     public String editUser (@PathVariable("id") Integer id, Model model){
         model.addAttribute("user", userService.getUserByID(id));
-        model.addAttribute("users", userService.getAllUsers());
+        model.addAttribute("users", userService.getTeamMates(userService.getCurrentUser()));
 
         return "adminPageTest";
+    }
+
+    private String getPrincipal(){
+        String userName = null;
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (principal instanceof UserDetails) {
+            userName = ((UserDetails)principal).getUsername();
+        } else {
+            userName = principal.toString();
+        }
+        return userName;
     }
 }
